refactor: 리팩토링 (#201)
