+++ src/main/java/com/api/ttoklip/domain/newsletter/post/repository/NewsletterQueryDslRepository.java
@@ -21,5 +21,15 @@ public interface NewsletterQueryDslRepository {
 
     List<Newsletter> getRecent3();
 
+    List<Newsletter> getHouseWorkNewsletter10Desc();
+
+    List<Newsletter> getRecipeNewsletter10Desc();
+
+    List<Newsletter> getSafeLivingNewsletter10Desc();
+
+    List<Newsletter> getWelfarePolicyNewsletter10Desc();
+
     List<Newsletter> findRandom4ActiveNewsletters();
+
+    Page<Newsletter> getContain(final String keyword, final Pageable pageable, final String sort);
 }
