+++ src/main/java/com/api/ttoklip/domain/search/service/SearchService.java
@@ -3,7 +3,7 @@ package com.api.ttoklip.domain.search.service;
 import com.api.ttoklip.domain.honeytip.post.domain.HoneyTip;
 import com.api.ttoklip.domain.honeytip.post.repository.HoneyTipSearchRepository;
 import com.api.ttoklip.domain.newsletter.post.domain.Newsletter;
-import com.api.ttoklip.domain.newsletter.post.repository.NewsletterSearchRepository;
+import com.api.ttoklip.domain.newsletter.post.repository.NewsletterQueryDslRepository;
 import com.api.ttoklip.domain.search.response.CommunityPaging;
 import com.api.ttoklip.domain.search.response.CommunitySingleResponse;
 import com.api.ttoklip.domain.search.response.HoneyTipPaging;
@@ -24,8 +24,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class SearchService {
 
     private final HoneyTipSearchRepository honeyTipSearchRepository;
-    private final NewsletterSearchRepository newsletterSearchRepository;
     private final CommunitySearchRepository communitySearchRepository;
+    private final NewsletterQueryDslRepository newsletterQueryDslRepository;
 
     public HoneyTipPaging honeyTipSearch(final String keyword, final Pageable pageable, final String sort) {
         Page<HoneyTip> contentPaging = honeyTipSearchRepository.getContain(keyword, pageable, sort);
@@ -48,7 +48,7 @@ public class SearchService {
     }
 
     public NewsletterPaging newsletterPaging(final String keyword, final Pageable pageable, final String sort) {
-        Page<Newsletter> contentPaging = newsletterSearchRepository.getContain(keyword, pageable, sort);
+        Page<Newsletter> contentPaging = newsletterQueryDslRepository.getContain(keyword, pageable, sort);
 
         // List<Entity>
         List<Newsletter> contents = contentPaging.getContent();
